Fix for issue delphix#49 - Make changes in JDBC Drivers POST/PUT API call to use file upload endpoint to restore JDBC driver
Implementation Details:
1- Recently we changed the JDBC Drivers endpoint in Masking Application to use file upload endpoint so it needs to change API calling in masking-initializer as well.
2- Also, JDBC Driver's property loggerInstalled also needs to add for phone-home.
